requirements.txt
langchain
pandas
chromadb
streamlit
pyyaml